requirements.txt
torch>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
streamlit>=1.30.0
pillow>=10.0.0
numpy>=1.23.0
sahi
roboflow
gdown
